+++ packages/datadog_flutter_plugin/android/src/main/kotlin/com/datadoghq/flutter/DatadogSdkPlugin.kt
@@ -23,6 +23,7 @@ import java.util.concurrent.atomic.AtomicBoolean
 class DatadogSdkPlugin : FlutterPlugin, MethodCallHandler {
     companion object ErrorCodes {
         const val CONTRACT_VIOLATION = "DatadogSdk:ContractViolation"
+        const val INVALID_OPERATION = "DatadogSdk:InvalidOperation"
     }
 
     private lateinit var channel: MethodChannel
@@ -102,19 +103,17 @@ class DatadogSdkPlugin : FlutterPlugin, MethodCallHandler {
             config.trackingConsent
         )
 
-        if (config.loggingConfiguration != null) {
-            logsPlugin = DatadogLogsPlugin()
-            logsPlugin?.setup(binding, config.loggingConfiguration!!)
-        }
+        // Always setup logging as a user can create a log after initialization
+        logsPlugin = DatadogLogsPlugin().apply { setup(binding) }
 
         if (config.tracingConfiguration != null) {
-            tracesPlugin = DatadogTracesPlugin()
-            tracesPlugin?.setup(binding, config.tracingConfiguration!!)
+            tracesPlugin = DatadogTracesPlugin().apply {
+                setup(binding, config.tracingConfiguration!!)
+            }
         }
 
         if (config.rumConfiguration != null) {
-            rumPlugin = DatadogRumPlugin()
-            rumPlugin?.setup(binding, config.rumConfiguration!!)
+            rumPlugin = DatadogRumPlugin().apply { setup(binding, config.rumConfiguration!!) }
         }
     }