+++ packages/datadog_flutter_plugin/android/src/main/kotlin/com/datadoghq/flutter/Helpers.kt
@@ -9,3 +9,11 @@ fun MethodChannel.Result.missingParameter(methodName: String, details: Object? =
         details
     )
 }
+
+fun MethodChannel.Result.invalidOperation(message: String, details: Object? = null) {
+    this.error(
+        DatadogSdkPlugin.INVALID_OPERATION,
+        message,
+        details
+    )
+}