+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Cancel spans on DatadogTrackingHttpClient when RUM is enabled (prevent spans
   from leaking native resources)
 * Remove native view tracking (Activities and Fragments) from Android by default
+* Add support for creating multiple named loggers: `DatadogSdk.createLogger` and
+  `LoggingConfiguration.loggerName`
+* Add support for configuring whether loggers send data to Datadog:
+  `LoggingConfiguration.sendLogsToDatadog`
 
 ## 1.0.0-alpha.1